Allow building with g++ as C compiler
  This change fixes the build errors and finishes with gcc 13, with several
warnings, on x86_64.
  It will not build with disassembler enabled unless extra customizations
are done.
  Besides working, for x86_64 after this commit, Lightning should not be
built with a C++ compiler.

+++ lib/jit_x86-cpu.c
@@ -50,8 +50,8 @@
 #    define stxr(u, v, w)		stxr_l(u, v, w)
 #    define stxi(u, v, w)		stxi_l(u, v, w)
 #    define can_sign_extend_int_p(im)					\
-	(((im) >= 0 && (long long)(im) <=  0x7fffffffLL) ||		\
-	 ((im) <  0 && (long long)(im) >  -0x80000000LL))
+	(((long long)(im) >= 0 && (long long)(im) <=  0x7fffffffLL) ||	\
+	 ((long long)(im) <  0 && (long long)(im) >  -0x80000000LL))
 #    define can_zero_extend_int_p(im)					\
 	((im) >= 0 && (im) < 0x80000000LL)
 #    define fits_uint32_p(im)		(((im) & 0xffffffff00000000LL) == 0)
